Allow 'some-type:main' modules to be looked up easier.
This removes the need to define custom resolver methods for every type
that also has a `main` module.

Specifically, in Ember-land we have at least two of these (possibly
more):

* `adapter:main`
* `router:main`

Prior to this change you would have to add a `resolveRouter` method
to the resolver. This instructs the resolver to update look for a module
with the `prefix + type`, and falls back on the existing logic if it is
not found.