Provide fallback lookup for partials.
Previously, a file named `blamo/snork/_blah-zorz` would not be able to
be looked up. This is problematic since Ember looks up partials with the
last part of the name prefixed with an underscore. This is then
normalized by our custom `normalize` function before lookup (into
`blamo/snork/-blah-zorz'), thus causing a miss on our filename.

A deprecation notice will be logged explaining the issue (and how to
correct it).