Improve performance of translateToContainerFullname function
While CPU profiling our app, we noticed that the
translateToContainerFullname function was taking quite a lot of time,
and this seemed to be mostly on account of regex use. This commit replaces
the regexes with equivalent string manipulation logic. The
translateToContainerFullname disappeared from the offenders list in our
CPU profiles after the change.