Move loader.js into subdirectory.
This is not technically needed, but it makes life a little easier when
addons have bad `treeForVendor` hooks that include all of
`node_modules`.

When that happens, we get an error saying that `loader.js` is both a
file (from our `lib` folder before this change) and a directory.(from
our own folder in `node_modules` for this package).

This avoids the issue by always using a subfolder for the imported file.