Do not put modules in pending list if they have been finalized
With the 2 pass approach to finding deps and evaluating them squentially, we are putting modules that have already been evaluated as well. This is not needed since when the module is read from the pending queue, it will return. However, we do not really need to push it in the pending queue and then know if it was finalized.

This is a minor optimization to the pending queue.