Don't resolve camelCase helper names to dasherized modules
In an Ember template, `{{someExpression}}` should NOT be
rendered using a `some-expression` helper. Prior to this commit,
it was. This led to a confusing experience where a local property
was not rendered because a dasherized helper of a similar name
existed.

This commit introduces a warning in debug mode when a camelCase
helper is looked up where the old behavior would have resulted
in a dash-named module being returned but no longer does.

This warning should be removed in a future major release of this
package.