Fix MU for deeply nested classic components
In a MU app consuming a classic addon that exposes nested components, they seem to work as long there is just one "layer" below, like components/foo/bar.js, but not for deeper layered components, like components/foo/bar/baz.js. This fails when resolving, it somehow thinks bar is a collection, and tries to get the default type from undefined. This throws in the @glimmer/resolver, so before the fallback is even tried.