Introduce ambient type definitions
Ultimately we'd like to convert the package to TS, but this unblocks
wider ecosystem adoption of the stable/preview Ember types. This PR can
also serve as a model for other packages which want to publish types in
a similar way.

This will also work just fine with the `@types` packages from
DefinitelyTyped once [a key PR][pr] lands there, but that is not the
path we will recommend, so this only includes the `ember-source` peer
dependency.

[pr]: https://github.com/DefinitelyTyped/DefinitelyTyped/pull/63589

(cherry picked from commit 3907d48de6511cc89fcb4a650df68b65d3c1efa9)

# Conflicts:
#	addon/package.json
#	package-lock.json
#	test-app/package.json